app/requirements.txt
django-rq==2.4.0
djangorestframework==3.12.2
whitenoise==5.2.0
requests==2.25.1
lxml==4.6.2
dateparser==1.0.0
gunicorn==20.0.4
drf-yasg[validation]
